Learned how to catch errors
I was given a template that turns a specified block that a player right clicks into a diamond block. I created a try and catch to catch the error because an error happens when you right click air (it's null). I played around with different ways to catch errors but this is the cleanest code I have for this example (and I understand it the best so far).