Learned about arrays
Used an example of arrays to list a players inventory contents. I made it so if it is empty it will give you a diamond, OR if you have a diamond in your inventory it'll add to your stack. I know there are more optimized methods for this exact feature but it's what I learned with.